Feat(storage): 强类型化存储引用仓储契约

+++ kuzhambu-servers/biz/storage/kuzhambu-storage-infra/src/main/java/com/thundax/kuzhambu/storage/infra/object/repository/impl/StoredObjectReferenceRepositoryImpl.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.thundax.kuzhambu.storage.domain.object.codec.StoredObjectIdCodec;
 import com.thundax.kuzhambu.storage.domain.object.model.entity.StoredObject;
 import com.thundax.kuzhambu.storage.domain.object.model.entity.StoredObjectReference;
+import com.thundax.kuzhambu.storage.domain.object.model.enums.StorageOwnerType;
 import com.thundax.kuzhambu.storage.domain.object.model.valueobject.StorageOwnerRef;
 import com.thundax.kuzhambu.storage.domain.object.model.valueobject.StoredObjectId;
 import com.thundax.kuzhambu.storage.domain.object.repository.StoredObjectReferenceRepository;
@@ -27,7 +28,7 @@ public class StoredObjectReferenceRepositoryImpl implements StoredObjectReferenc
     }
 
     @Override
-    public List<String> listReferenceOwnerTypes() {
+    public List<StorageOwnerType> listReferenceOwnerTypes() {
         return mapper
                 .selectObjs(new QueryWrapper<StoredObjectReferenceDO>()
                         .select("reference_owner_type")
@@ -36,6 +37,7 @@ public class StoredObjectReferenceRepositoryImpl implements StoredObjectReferenc
                 .stream()
                 .filter(Objects::nonNull)
                 .map(String::valueOf)
+                .map(StorageOwnerType::from)
                 .collect(Collectors.toList());
     }
 
@@ -119,9 +121,12 @@ public class StoredObjectReferenceRepositoryImpl implements StoredObjectReferenc
     }
 
     @Override
-    public void deleteByObjectId(String id) {
+    public void deleteByObjectId(StoredObjectId objectId) {
+        if (objectId == null) {
+            return;
+        }
         LambdaQueryWrapper<StoredObjectReferenceDO> wrapper = new LambdaQueryWrapper<>();
-        wrapper.eq(StoredObjectReferenceDO::getObjectId, Long.valueOf(id));
+        wrapper.eq(StoredObjectReferenceDO::getObjectId, objectId.value());
         mapper.delete(wrapper);
     }
